fix(landing): hide package releases from changelog
